Move main components/reducers into seperate folder
