+++ Frontend/src/main/reducers/index.ts
@@ -1,10 +1,10 @@
 import reduceReducers from 'reduce-reducers';
 import { Action, combineReducers, Reducer } from 'redux';
 
-import { reducer as instructorReducer, rootReducer as instructorRootReducer } from '../instructor/reducers';
-import { InstructorState } from '../instructor/state';
-import { reducer as registrationReducer, rootReducer as registrationRootReducer } from '../registration/reducers';
-import { RegistrationState } from '../registration/state';
+import { reducer as instructorReducer, rootReducer as instructorRootReducer } from '../../instructor/reducers';
+import { InstructorState } from '../../instructor/state';
+import { reducer as registrationReducer, rootReducer as registrationRootReducer } from '../../registration/reducers';
+import { RegistrationState } from '../../registration/state';
 
 export interface ReducerTree {
     instructor: Reducer<InstructorState, Action>;
