+++ Frontend/src/main/state/index.ts
@@ -1,0 +1,7 @@
+import { InstructorState } from "../../instructor/state";
+import { RegistrationState } from "../../registration/state";
+
+export interface State {
+    instructor: InstructorState;
+    registration: RegistrationState;
+}
