+++ Frontend/src/registration/containers/RegistrationStep2Container.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 
-import { State } from '../../state';
+import { State } from '../../main/state';
 import { applicantChange, selectCoursesForParticipants, loadPossibleCourses } from '../actions';
 import { ApplicantProps } from '../components/Applicant';
 import { CourseSelection, CourseSelectionProps, SelectionMap } from '../components/CourseSelection';
