--- Frontend/src/state/index.ts
@@ -1,7 +1,0 @@
-import { RegistrationState } from "../registration/state";
-import { InstructorState } from "../instructor/state";
-
-export interface State {
-    instructor: InstructorState;
-    registration: RegistrationState;
-}
